docs: ✏️ Add documentation
